requirements.txt
Package                   Version
------------------------- -----------
aiohappyeyeballs          2.4.0
aiohttp                   3.10.5
aiosignal                 1.3.1
altair                    5.4.0
annotated-types           0.7.0
anyio                     4.4.0
argcomplete               3.4.0
async-timeout             4.0.3
attrs                     24.2.0
blinker                   1.8.2
cachetools                5.5.0
certifi                   2024.7.4
charset-normalizer        3.3.2
click                     8.1.7
distro                    1.9.0
exceptiongroup            1.2.2
fastapi                   0.112.1
Flask                     3.0.3
frozenlist                1.4.1
gitdb                     4.0.11
GitPython                 3.1.43
h11                       0.14.0
httpcore                  1.0.5
httptools                 0.6.1
httpx                     0.27.0
idna                      3.7
itsdangerous              2.2.0
Jinja2                    3.1.4
jiter                     0.5.0
jsonschema                4.23.0
jsonschema-specifications 2023.12.1
markdown-it-py            3.0.0
MarkupSafe                2.1.5
mdurl                     0.1.2
multidict                 6.0.5
narwhals                  1.5.4
numpy                     2.1.0
openai                    1.42.0
packaging                 24.1
pandas                    2.2.2
pillow                    10.4.0
pip                       24.2
pipx                      1.6.0
platformdirs              4.2.2
protobuf                  5.27.3
pyarrow                   17.0.0
pydantic                  2.8.2
pydantic_core             2.20.1
pydeck                    0.9.1
Pygments                  2.18.0
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
pytz                      2024.1
PyYAML                    6.0.2
referencing               0.35.1
requests                  2.32.3
rich                      13.7.1
rpds-py                   0.20.0
setuptools                72.2.0
six                       1.16.0
smmap                     5.0.1
sniffio                   1.3.1
starlette                 0.38.2
streamlit                 1.37.1
tenacity                  8.5.0
toml                      0.10.2
tomli                     2.0.1
tornado                   6.4.1
tqdm                      4.66.5
typing_extensions         4.12.2
tzdata                    2024.1
urllib3                   2.2.2
userpath                  1.9.2
uvicorn                   0.30.6
uvloop                    0.20.0
watchdog                  4.0.2
watchfiles                0.23.0
websockets                13.0
Werkzeug                  3.0.4
wheel                     0.44.0
yarl                      1.9.4
requests